Refactored code into UrlBuilder util. Added summary to some utils.
